refactor(analyzer): extract call-graph type model into call-graph-types.ts (slice 1/5)
First safe slice of modularize-call-graph-builder, behind the stable barrel. Moves the
TYPES section (edge/node/class model, CallGraphResult/SerializedCallGraph, CALL_DISTANCE_COSTS,
callDistance, layerOf, classifyLayerEdge) out of the 5,425-line call-graph.ts into a new
dependency-light call-graph-types.ts. call-graph.ts re-exports every public name, so none of
the 155 importers move; RawEdge and CALL_DISTANCE_FALLBACK stay internal (off the surface).
call-graph.ts: 5,425 → 5,150 lines.

Behavior-preserving, verified four ways:
- Byte-level snapshot oracle (multi-language CallGraphBuilder.build() + the moved pure helpers,
  serialized canonically + SHA-256'd) hashes IDENTICALLY before/after: 131ba4c6…
- Public export surface byte-for-byte identical (multi-line-aware diff of call-graph.ts exports).
- tsc compiles all 155 importers clean; build/lint/typecheck green.
- Full suite green: 279 files / 5531 tests.

New `stable call-graph barrel` test locks the invariant (re-export identity === source binding +
moved-helper behavior) so the remaining slices (nodes/extract/dispatch/grammar-loader) can't
silently break the surface. No feature, no dependency, no LLM, no persisted artifact.
